changed table design and reading code

+++ bff/internal/model/merchant.go
@@ -2,6 +2,6 @@ package model
 
 type Merchant struct {
 	Name     string   `json:"name"`
-	Amount   float64  `json:"amount"`
+	Amount   int      `json:"amount"`
 	Category Category `json:"category"`
 }
